refactor(sessions): move StatusDot into sessions/components
Phase 4 of 4 toward issue #181. Move StatusDot (with test) from
workspace/components/ to sessions/components/. The component takes
SessionStatus and is consumed only by Card and Tab — both in
sessions/ — so it belongs there.

After this phase:
  - src/features/sessions/ is the self-contained session-domain feature
    (types/, hooks/, utils/, components/) with zero upward dependency
    into workspace/.
  - src/features/workspace/sessions/ no longer exists.
  - workspace/types/ retains shell types (NavigationItem, Project,
    ContextPanelType, ContextPanelState, Terminal, WorkspaceState);
    Project.sessions: Session[] imports Session from sessions/types.
  - agents/registry.ts stays at src/agents/ (cross-cutting registry).
  - SidebarStatusHeader and useNotifyInfo stay in workspace/ (no
    session-domain coupling).

Closes #181
