+++ src/features/sessions/components/Card.tsx
@@ -1,7 +1,7 @@
 import { type ReactElement } from 'react'
 import { Reorder } from 'framer-motion'
 import type { Session } from '../types'
-import { StatusDot } from '../../workspace/components/StatusDot'
+import { StatusDot } from './StatusDot'
 import { useRenameState } from '../hooks/useRenameState'
 import { formatRelativeTime } from '../../agent-status/utils/relativeTime'
 import {
