+++ src/features/sessions/components/StatusDot.test.tsx
